cambio int por float

+++ src/CharactersServer/CharacterServer.h
@@ -41,7 +41,7 @@ public:
     void update(int distance, int posContrincante, actions_t actionRecieved);
     void positionUpdate(int *x);
     void startIntro();
-    virtual void makeBuilderStruct(character_builder_t *builder, bool firstTeam, int pos) = 0;
+    virtual void makeBuilderStruct(character_builder_t *builder, bool firstTeam, double pos) = 0;
     void makeUpdaterStruct(character_updater_t* updater);
     virtual int getSpriteNumber() = 0;
     actions_t getCurrentAction();
